+++ packages/kit/src/core/config/index.spec.js
@@ -73,6 +73,9 @@ const get_defaults = (prefix = '') => ({
 			reportOnly: directive_defaults
 		},
 		endpointExtensions: undefined,
+		env: {
+			publicPrefix: 'PUBLIC_'
+		},
 		files: {
 			assets: join(prefix, 'static'),
 			hooks: join(prefix, 'src/hooks'),
@@ -112,6 +115,7 @@ const get_defaults = (prefix = '') => ({
 			entries: ['*'],
 			force: undefined,
 			onError: 'fail',
+			origin: 'http://sveltekit-prerender',
 			pages: undefined
 		},
 		protocol: undefined,