+++ packages/kit/test/prerendering/basics/src/routes/origin/index.svelte
@@ -1,0 +1,9 @@
+<script>
+	import { page } from '$app/stores';
+
+	/** @type {string} */
+	export let message;
+</script>
+
+<h1>{message}</h1>
+<h2>{$page.url.origin}</h2>